+++ lib/liquid_kit.dart
@@ -5,4 +5,5 @@ export 'src/liquid_glass_navigation_bar.dart';
 export 'src/liquid_glass_physics.dart';
 export 'src/liquid_glass_theme.dart';
 export 'src/liquid_glass_search_bar.dart';
-
+export 'src/liquid_glass_sheet.dart';
+export 'src/liquid_glass_toolbar.dart';